Add interactive weight chart with range selection
Introduces WeightChartView for visualizing user weight history and integrates it into UserWeightFragment. Adds date range selection buttons to filter chart data (all, year, 6 months, 3 months, month, week) and updates related logic. Refactors weight handling to use Double instead of Int for improved precision. Updates layout to support the new chart and range buttons.

+++ app/src/main/java/com/example/fithub/data/UpdateUserDto.kt
@@ -7,7 +7,7 @@ data class UpdateUserDto(
 )
 
 data class UpdateProfileData(
-    val weightKg: Int? = null,
+    val weightKg: Double? = null,
     val heightCm: Int? = null,
     val sex: String? = null,
     val birthDate: String? = null
